backend/requirements.txt
fastapi
uvicorn
biopython
requests
python-dotenv
openai